# MF0966_3_KUARENTENA/docs/mf0966_3_swagger.yaml
openapi: 3.0.1
info:
  title: MF0966_3
  description: Swagger para el examen MF0966_3 version Cuarentena
  contact:
    email: alainzaba@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/mf0966_3/api/'
paths:
  /resenyas:
    get:
      tags:
        - Reseñas
      summary: Obtener todos las reseñas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resenya'
    post:
      tags:
      - Reseñas
      summary: Dar de alta una reseña nueva
      requestBody:
        description: Reseña a añadir
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenya'
        required: true
      responses:
        '201':
          description: Reseña creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resenya'
  /resenyas/{idAlumno}/{idCurso}:
    get:
      tags:
        - Reseñas
      summary: Obtener cliente por id
      parameters:
        - name: idAlumno
          in: path
          description: ID del alumno
          required: true
          schema:
            type: integer
            format: int64
        - name: idCurso
          in: path
          description: ID del curso
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resenya'
        '404':
          description: No se ha encontrado la Reseña
    put:
      tags:
        - Reseñas
      summary: Modificar un cliente existente
      parameters:
        - name: idAlumno
          in: path
          description: ID del alumno
          required: true
          schema:
            type: integer
            format: int64
        - name: idCurso
          in: path
          description: ID del curso
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Cliente a modificar
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenya'
        required: true
      responses:
        '200':
          description: Reseña modificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resenya'
        '400':
          description: No concuerdan los ids
        '404':
          description: No se ha encontrado la Reseña
        
    delete:
      tags:
        - Reseñas
      summary: Borrar una Reseña por idAlumno e idCurso
      parameters:
        - name: idAlumno
          in: path
          description: ID del alumno
          required: true
          schema:
            type: integer
            format: int64
        - name: idCurso
          in: path
          description: ID del curso
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reseña borrado
        '404':
          description: No se ha encontrado la Reseña a eliminar
  /cursos:
    get:
      tags:
        - Cursos
      summary: Obtener todos los Cursos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Curso'
  /cursos/{id}/:
    get:
      tags:
        - Cursos
      summary: Obtener Curso por id
      parameters:
        - name: id
          in: path
          description: ID del curso
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curso'
        '404':
          description: No se ha encontrado el Curso
components:
  schemas:
    Alumno:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        nombre:
          type: "string"
        apellidos:
          type: "string"
    Profesor:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        nombre:
          type: "string"
        apellidos:
          type: "string"
    Curso:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        nombre:
          type: "string"
        nHoras:
          type: "integer"
          format: "int64"
        profesor:
          $ref: '#/components/schemas/Profesor'
        resenyas:
          type: "array"
          items:
            $ref: '#/components/schemas/Resenya'
    Resenya:
      type: "object"
      properties:
        alumno:
          $ref: '#/components/schemas/Alumno'
        curso:
          type: "object"
          properties:
            id:
              type: "integer"
              format: "int64"
            nombre:
              type: "string"
            nHoras:
              type: "integer"
              format: "int64"
            profesor:
              $ref: '#/components/schemas/Profesor'
        resenya:
          type: "string"
        created_at:
          type: "string"
        updated_at: 
          type: "string"